rename package to exception
