+++ src/main/java/com/softserve/edu/hypercinema/exception/OrderException.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.hypercinema.customexception;
+package com.softserve.edu.hypercinema.exception;
 
 public class OrderException extends RuntimeException {
     public OrderException() {
